working theme switcher over local storage
