+++ apps/designer/app/shared/design-system/index.ts
@@ -11,7 +11,6 @@ export {
   theme,
   type CSS,
 } from "./stitches.config";
-export * from "./components/theme";
 
 export * as Collapsible from "@radix-ui/react-collapsible";
 export { AccessibleIcon } from "@radix-ui/react-accessible-icon";
